Drop yt/ dq relocation; keep ydb-only provider cleanup.
Revert move_dq_config changes under yt/ and restore ydb layout while keeping PEERDIR tidy-ups and DqStrippedSuffied helper.
