+++ ydb/library/yql/providers/dq/provider/yql_dq_control.h
@@ -41,4 +41,6 @@ IDqControlFactoryPtr CreateDqControlFactory(
     const TMap<TString, TString>& udfs,
     const TFileStoragePtr& fileStorage);
 
+extern const TString DqStrippedSuffied;
+
 } // namespace NYql
