+++ ydb/library/yql/providers/dq/provider/yql_dq_gateway.cpp
@@ -1,6 +1,4 @@
-#include "yql_dq_gateway_factory.h"
-
-#include <yt/yql/providers/dq/config/config.pb.h>
+#include "yql_dq_gateway.h"
 
 #include <yql/essentials/providers/common/provider/yql_provider_names.h>
 #include <ydb/library/yql/providers/dq/api/grpc/api.grpc.pb.h>
@@ -9,6 +7,7 @@
 #include <yql/essentials/utils/backtrace/backtrace.h>
 #include <yql/essentials/utils/failure_injector/failure_injector.h>
 #include <yql/essentials/public/issue/yql_issue_message.h>
+#include <ydb/library/yql/providers/dq/config/config.pb.h>
 #include <yql/essentials/utils/log/log.h>
 
 #include <ydb/public/lib/yson_value/ydb_yson_value.h>
