+++ ydb/library/yql/providers/dq/service/ya.make
@@ -26,7 +26,6 @@ PEERDIR(
     ydb/library/yql/providers/dq/interface
     ydb/library/yql/providers/dq/worker_manager
     ydb/library/yql/providers/dq/worker_manager/interface
-    yt/yql/providers/dq/config
 )
 
 YQL_LAST_ABI_VERSION()
