+++ ydb/library/yql/providers/dq/ya.make
@@ -10,6 +10,8 @@ RECURSE(
     planner
     provider
     runtime
+    service
+    stats_collector
     task_runner
     task_runner_actor
     worker_manager
