+++ ydb/library/yql/tools/dqrun/dqrun.cpp
@@ -1,4 +1,4 @@
-#include <yt/yql/tools/dqrun/lib/dqrun_lib.h>
+#include <ydb/library/yql/tools/dqrun/lib/dqrun_lib.h>
 #include <util/generic/yexception.h>
 
 int main(int argc, const char *argv[]) {
