+++ ydb/library/yql/tools/dqrun/examples/fs.conf
