+++ ydb/library/yql/tools/dqrun/examples/gateways.conf
