+++ ydb/library/yql/tools/dqrun/lib/dqrun_lib.cpp
@@ -38,8 +38,7 @@
 #include <ydb/library/yql/providers/generic/provider/yql_generic_provider.h>
 #include <ydb/library/yql/providers/generic/actors/yql_generic_provider_factories.h>
 #include <ydb/library/yql/providers/dq/interface/yql_dq_task_preprocessor.h>
-#include <yt/yql/providers/dq/gateway/yql_dq_gateway_factory.h>
-#include <yt/yql/providers/dq/local_gateway/yql_dq_gateway_local.h>
+#include <ydb/library/yql/providers/dq/local_gateway/yql_dq_gateway_local.h>
 #include <ydb/library/yql/providers/dq/provider/exec/yql_dq_exectransformer.h>
 #include <ydb/library/yql/providers/dq/provider/yql_dq_provider.h>
 #include <ydb/library/yql/providers/dq/helper/yql_dq_helper_impl.h>
