+++ ydb/library/yql/tools/dqrun/ya.make
@@ -22,20 +22,23 @@ IF (NOT OS_WINDOWS)
     )
 
     PEERDIR(
-        library/cpp/lfalloc/alloc_profiler
-        ydb/library/yql/dq/comp_nodes/llvm16
-        ydb/library/yql/providers/pq/gateway/dummy
-        ydb/library/yql/udfs/common/clickhouse/client
-        ydb/public/sdk/cpp/src/client/persqueue_public/codecs
-        yql/essentials/minikql/comp_nodes/llvm16
+        ydb/library/yql/tools/dqrun/lib
+
+        yt/yql/providers/yt/codec/codegen
+        yt/yql/providers/yt/comp_nodes/llvm16
+        yt/yql/providers/yt/comp_nodes/dq/llvm16
         yql/essentials/minikql/invoke_builtins/llvm16
+        yql/essentials/minikql/comp_nodes/llvm16
         yql/essentials/parser/pg_wrapper
         yql/essentials/public/udf/service/exception_policy
         yql/essentials/sql/pg
-        yt/yql/providers/yt/codec/codegen
-        yt/yql/providers/yt/comp_nodes/dq/llvm16
-        yt/yql/providers/yt/comp_nodes/llvm16
-        yt/yql/tools/dqrun/lib
+
+        library/cpp/lfalloc/alloc_profiler
+
+        ydb/library/yql/udfs/common/clickhouse/client
+        ydb/library/yql/dq/comp_nodes/llvm16
+        ydb/library/yql/providers/pq/gateway/dummy
+        ydb/public/sdk/cpp/src/client/persqueue_public/codecs
     )
 
     YQL_LAST_ABI_VERSION()
