+++ ydb/library/yql/tools/ya.make
@@ -1,3 +1,4 @@
 RECURSE(
+    dqrun
     solomon_emulator
 )
