--- yt/yql/providers/dq/gateway/yql_dq_gateway_factory.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include <ydb/library/yql/providers/dq/provider/yql_dq_gateway.h>
-
-namespace NYql {
-
-namespace NProto {
-class TDqConfig;
-}
-
-TIntrusivePtr<IDqGateway> CreateDqGateway(const TString& host, int port);
-TIntrusivePtr<IDqGateway> CreateDqGateway(const NProto::TDqConfig& config);
-
-} // namespace NYql
