--- yt/yql/providers/dq/provider/ya.make
@@ -1,3 +1,0 @@
-RECURSE_FOR_TESTS(
-    ut
-)
